==> 02_ScrapeEnglishIdioms.ipynb <==
# English Idiom Scraping

Grab full list of English idions from Wikitionary.

# Collected for our paper based on Wikitionary version on 2022-02-1

from bs4 import BeautifulSoup
import requests
import time

wiki_base = "https://en.wiktionary.org"
idiom_page = "/wiki/Category:English_idioms"

soup = BeautifulSoup(requests.get(wiki_base + idiom_page).content, 'html.parser')
next_link = soup.find_all("a", text="next page")[0]["href"]

category_divs = []
while next_link:
    category_divs.append(soup.find_all("div", class_="mw-category"))
    print("Fetching " + wiki_base + next_link)
    soup = BeautifulSoup(requests.get(wiki_base + next_link).content, 'html.parser')
    
    next_link_el = soup.find_all("a", text="next page")
    if len(next_link_el) > 0:
        next_link = next_link_el[0]["href"]
    else:
        next_link = None
        
    time.sleep(1)

category_divs.append(soup.find_all("div", class_="mw-category"))

parsed = []
for d in category_divs:
    for a in d[0].find_all("a"):
        parsed.append(a.contents[0])

with open("english_idioms.txt", "w") as f:
    for p in parsed:
        f.write(p + "\n")
